folder structure reorg for climacore
