+++ src/ClimaAtmos.jl
@@ -2,49 +2,15 @@ module ClimaAtmos
 
 using Pkg, LinearAlgebra
 
-# convenience functions
-export @boilerplate
-
 # example 
 export function_documentation_template
 
 # include other modules
 include("utils/Utils.jl")
 include("interface/Interface.jl")
-include("backends/backends.jl")
-
-
-# convenience functions: definitions
-"""
-macro boilerplate()
-
-# Description 
-A convenience function that includes all the usual needed for a script
-"""
-macro boilerplate()
-    boiler_block = :( 
-        using ClimaAtmos.Utils;
-        using ClimaAtmos.Interface;
-        using ClimaAtmos.Backends;
-        using IntervalSets, UnPack;
-        using OrdinaryDiffEq: SSPRK33;
-
-        # explicit imports from Interface
-        import ClimaAtmos.Interface: PeriodicRectangle, SingleColumn;
-        import ClimaAtmos.Interface: BarotropicFluidModel, HydrostaticModel;
-        import ClimaAtmos.Interface: TimeStepper;
-        import ClimaAtmos.Interface: DirichletBC, Simulation;
-        import ClimaAtmos.Backends: create_ode_problem, evolve;
-
-        # imports from Clima Core
-        using ClimaCore.Geometry;
-        import ClimaCore: Fields, Domains, Topologies, Meshes, Spaces
-        )
-    return boiler_block
-end
+include("backends/Backends.jl")
 
 # example: definition
-
 """
 function_documentation_template(a; informative_keyword = "yes")
