+++ src/interface/simulations.jl
@@ -1,5 +1,8 @@
 abstract type AbstractSimulation end
 
+"""
+    Simulation <: AbstractSimulation
+"""
 struct Simulation{A,B,C,D,E} <: AbstractSimulation
     backend::A
     model::B
